mainForm

+++ clinicapp/src/main/java/com/tesis/clinicapp/model/Citas.java
@@ -19,7 +19,7 @@ import javax.persistence.TemporalType;
 @Table(name = "citas", schema = "clinica")
 public class Citas implements java.io.Serializable {
 
-	private int id;
+	private Long id;
 	private Paciente paciente;
 	private Date fechaReserva;
 	private String descripcion;
@@ -27,7 +27,7 @@ public class Citas implements java.io.Serializable {
 	public Citas() {
 	}
 
-	public Citas(int id, Paciente paciente, Date fechaReserva, String descripcion) {
+	public Citas(Long id, Paciente paciente, Date fechaReserva, String descripcion) {
 		this.id = id;
 		this.paciente = paciente;
 		this.fechaReserva = fechaReserva;
@@ -37,11 +37,11 @@ public class Citas implements java.io.Serializable {
 	@Id
 
 	@Column(name = "id", unique = true, nullable = false)
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
