+++ clinicapp/src/main/java/com/tesis/clinicapp/model/Examen.java
@@ -22,7 +22,7 @@ import javax.persistence.TemporalType;
 @Table(name = "examen", schema = "clinica")
 public class Examen implements java.io.Serializable {
 
-	private int id;
+	private Long id;
 	private CatalogoExamen catalogoExamen;
 	private Laboratorista laboratorista;
 	private Paciente paciente;
@@ -33,14 +33,14 @@ public class Examen implements java.io.Serializable {
 	public Examen() {
 	}
 
-	public Examen(int id, CatalogoExamen catalogoExamen, Laboratorista laboratorista, Paciente paciente) {
+	public Examen(Long id, CatalogoExamen catalogoExamen, Laboratorista laboratorista, Paciente paciente) {
 		this.id = id;
 		this.catalogoExamen = catalogoExamen;
 		this.laboratorista = laboratorista;
 		this.paciente = paciente;
 	}
 
-	public Examen(int id, CatalogoExamen catalogoExamen, Laboratorista laboratorista, Paciente paciente,
+	public Examen(Long id, CatalogoExamen catalogoExamen, Laboratorista laboratorista, Paciente paciente,
 			Date fechaProcesamiento, String observaciones, Set<ItemsExamen> itemsExamens) {
 		this.id = id;
 		this.catalogoExamen = catalogoExamen;
@@ -54,11 +54,11 @@ public class Examen implements java.io.Serializable {
 	@Id
 
 	@Column(name = "id", unique = true, nullable = false)
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
